# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ApproveSettleBody": {
        "properties": {
          "adjustments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RatingAdjustment"
            },
            "title": "Adjustments",
            "type": "array"
          }
        },
        "required": [
          "adjustments"
        ],
        "title": "ApproveSettleBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeaderboardResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/LeaderboardEntry"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "LeaderboardResponse",
        "type": "object"
      },
...
      "MyRatingEventsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RatingEventWithNames"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "MyRatingEventsResponse",
        "type": "object"
      },
...
      "MyRatingsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MyRatingSummary"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "MyRatingsResponse",
        "type": "object"
      },
...
      "PlayerRatingResponse": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PlayerRating"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "data"
        ],
        "title": "PlayerRatingResponse",
        "type": "object"
      },
...
      "RatingAdjustment": {
        "properties": {
          "initial_rating": {
            "maximum": 4000.0,
            "minimum": 0.0,
            "title": "Initial Rating",
            "type": "integer"
          },
          "player_rating_id": {
            "title": "Player Rating Id",
            "type": "integer"
          }
        },
        "required": [
          "player_rating_id",
          "initial_rating"
        ],
        "title": "RatingAdjustment",
        "type": "object"
      },
      "RatingCategoriesResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/RatingCategory"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "RatingCategoriesResponse",
        "type": "object"
      },
...
      "RatingCategoryCreateBody": {
        "properties": {
          "algorithm": {
            "default": "chinatt",
            "title": "Algorithm",
            "type": "string"
          },
          "key": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Key",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "name",
          "algorithm"
        ],
        "title": "RatingCategoryCreateBody",
        "type": "object"
      },
      "RatingCategoryResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RatingCategory"
          }
        },
        "required": [
          "data"
        ],
        "title": "RatingCategoryResponse",
        "type": "object"
      },
      "RatingCategoryUpdateBody": {
        "properties": {
          "algorithm": {
            "default": "chinatt",
            "title": "Algorithm",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "algorithm"
        ],
        "title": "RatingCategoryUpdateBody",
        "type": "object"
      },
...
      "SeedRatingBody": {
        "properties": {
          "rating": {
            "maximum": 4000.0,
            "minimum": 0.0,
            "title": "Rating",
            "type": "integer"
          }
        },
        "required": [
          "rating"
        ],
        "title": "SeedRatingBody",
        "type": "object"
      },
      "SettlementResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SettlementResult"
          }
        },
        "required": [
          "data"
        ],
        "title": "SettlementResponse",
        "type": "object"
      },
...
      "SettlementReviewsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/SettlementReviewItem"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "SettlementReviewsResponse",
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n### Description\nThis API allows you to do everything the frontend of [Bracket](https://github.com/evroon/bracket)\nallows you to do (the frontend uses this API as well).\n\nFore more information, see the [documentation](https://docs.bracketapp.nl).\n\n### Table of Contents\n*(links only work for [ReDoc](https://api.bracketapp.nl/redoc), not for Swagger UI)*\n\n- [Auth](#tag/Auth)\n\n- [Clubs](#tag/Clubs)\n\n- [Courts](#tag/Courts)\n\n- [Favorites](#tag/Favorites)\n\n- [Internals](#tag/Internals)\n\n- [Matches](#tag/Matches)\n\n- [Participants](#tag/Participants)\n\n- [Players](#tag/Players)\n\n- [Rankings](#tag/Rankings)\n\n- [Ratings](#tag/Ratings)\n\n- [Rounds](#tag/Rounds)\n\n- [Showcase](#tag/Showcase)\n\n- [Stage Items](#tag/Stage-Items)\n\n- [Stage Item Inputs](#tag/Stage-Item-Inputs)\n\n- [Stages](#tag/Stages)\n\n- [Teams](#tag/Teams)\n\n- [Tournaments](#tag/Tournaments)\n\n- [Users](#tag/Users)\n\n### Links\nGitHub: <https://github.com/evroon/bracket>\n\nDocs: <https://docs.bracketapp.nl>\n\nDemo: <https://www.bracketapp.nl/demo>\n\nAPI docs (Redoc): <https://api.bracketapp.nl/redoc>\n\nAPI docs (Swagger UI): <https://api.bracketapp.nl/docs>\n",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "summary": "API for Bracket, an open source tournament system.",
    "title": "Bracket API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/admin/settlement-reviews": {
      "get": {
        "operationId": "settlement_reviews_admin_settlement_reviews_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementReviewsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Settlement Reviews",
        "tags": [
          "Ratings"
        ]
      }
    },
    "/admin/settlements/{tournament_id}/approve-and-settle": {
      "post": {
        "description": "One admin action: approve every PENDING seed of this tournament's participants\n(applying any per-player adjustment) and settle \u2014 atomically. The adjusted seed\nbecomes the player's frozen entry rating, so the settlement reflects it.",
        "operationId": "approve_and_settle_admin_settlements__tournament_id__approve_and_settle_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveSettleBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Approve And Settle",
        "tags": [
          "Ratings"
        ]
      }
    },
...
    "/me/ratings": {
      "get": {
        "operationId": "my_ratings_me_ratings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRatingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "My Ratings",
        "tags": [
          "Ratings"
        ]
      }
    },
    "/me/ratings/{category_id}/events": {
      "get": {
        "operationId": "my_rating_events_me_ratings__category_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyRatingEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "My Rating Events",
        "tags": [
          "Ratings"
        ]
      }
    },
...
    "/rating-categories": {
      "get": {
        "operationId": "list_rating_categories_rating_categories_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingCategoriesResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Rating Categories",
        "tags": [
          "Ratings"
        ]
      },
      "post": {
        "operationId": "create_category_rating_categories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingCategoryCreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingCategoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Category",
        "tags": [
          "Ratings"
        ]
      }
    },
    "/rating-categories/{category_id}": {
      "delete": {
        "operationId": "remove_category_rating_categories__category_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remove Category",
        "tags": [
          "Ratings"
        ]
      },
      "put": {
        "operationId": "update_category_rating_categories__category_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RatingCategoryUpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RatingCategoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Category",
        "tags": [
          "Ratings"
        ]
      }
    },
    "/rating-categories/{category_id}/leaderboard": {
      "get": {
        "operationId": "leaderboard_rating_categories__category_id__leaderboard_get",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaderboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Leaderboard",
        "tags": [
          "Ratings"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/settle": {
      "post": {
        "operationId": "settle_tournaments__tournament_id__settle_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Settle",
        "tags": [
          "Ratings"
        ]
      }
    },
    "/tournaments/{tournament_id}/settlement-request": {
      "post": {
        "operationId": "settlement_request_tournaments__tournament_id__settlement_request_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Settlement Request",
        "tags": [
          "Ratings"
        ]
      }
    },
    "/tournaments/{tournament_id}/settlement-status": {
      "get": {
        "operationId": "settlement_status_tournaments__tournament_id__settlement_status_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettlementResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Settlement Status",
        "tags": [
          "Ratings"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/teams/{team_id}/seed-rating": {
      "put": {
        "operationId": "seed_team_rating_tournaments__tournament_id__teams__team_id__seed_rating_put",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedRatingBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRatingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Seed Team Rating",
        "tags": [
          "Ratings"
        ]
      }
    },
...
  }
}